+++ src/main/kotlin/blackjack/domain/Cards.kt
@@ -1,0 +1,10 @@
+package blackjack.domain
+
+import blackjack.domain.Rank.Companion.ACE
+
+data class Cards(val cards: List<Card>) {
+    fun calculateScore(): Int {
+        return cards.filterNot { it.rank == ACE }
+            .sumOf { it.rank.score }
+    }
+}
