+++ src/main/kotlin/blackjack/domain/Suit.kt
@@ -1,0 +1,8 @@
+package blackjack.domain
+
+enum class Suit {
+    SPADE,
+    HEART,
+    DIAMOND,
+    CLUB,
+}
