+++ src/test/kotlin/blackjack/BlackJackTest.kt
@@ -1,54 +1,15 @@
 package blackjack
 
-import blackjack.Rank.Companion.ACE
-import blackjack.Suit.SPADE
+import blackjack.domain.Card
+import blackjack.domain.Cards
+import blackjack.domain.Rank
+import blackjack.domain.Suit.SPADE
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
 
-enum class Suit {
-    SPADE,
-    HEART,
-    DIAMOND,
-    CLUB,
-}
-
-@JvmInline
-value class Rank(private val value: String) {
-    init {
-        require(value in RANK_VALUES) { "랭크는 $RANK_VALUES 에 포함되어야 합니다." }
-    }
-
-    val score: Int
-        get() =
-            when (value) {
-                in NUMBER_VALUES -> value.toInt()
-                in FACE_VALUES -> 10
-                else -> throw IllegalArgumentException("에이스는 경우에 따라 점수를 합산해야 합니다")
-            }
-
-    companion object {
-        private const val ACE_VALUE = "A"
-
-        private val NUMBER_VALUES = listOf("2", "3", "4", "5", "6", "7", "8", "9")
-        private val FACE_VALUES = listOf("J", "Q", "K")
-        private val RANK_VALUES = NUMBER_VALUES + FACE_VALUES + ACE_VALUE
-
-        val ACE = Rank(ACE_VALUE)
-    }
-}
-
-data class Card(val rank: Rank, val suit: Suit)
-
-data class Cards(val cards: List<Card>) {
-    fun calculateScore(): Int {
-        return cards.filterNot { it.rank == ACE }
-            .sumOf { it.rank.score }
-    }
-}
-
-class BlackjackTest : StringSpec({
+class BlackJackTest : StringSpec({
     "카드는 랭크와 문양으로 이루어진다" {
         val card = Card(Rank("3"), SPADE)
         card.rank shouldBe Rank("3")
